add "v" coordinate to patchCoords provided to basis curves ribbon surface shaders
Curve shaders that want to apply a texture to a curve being rendered as a
ribbon need some value that varies along both directions of a curve: one
along the length of the curve and one side-to-side on the width of the
ribbon. It's easy to surface a primvar that varies down the length of the
curve, but there's nothing that varies along the width.

Internal to the basisCurves.glslfx, though, we do have a "v" value that
varies along that axis, which is already used for procedurally generating
normals. Previously, this "v" value wasn't included in the patch
coordinates for basis curves, though; the patchCoords passed to the surface
shader was always a vec4 of zeroes. This adds that "v" value as the second
element in patchCoords so that surface shaders can use it.

Note that I don't add support for "u" here, even though the shader has
access to a "u" as well; that is because the "u" there is not very useful
for shading, because the value of "u" goes from zero to one over each small
quad patch generated for the ribbon, and thus doesn't really represent
anything inherent in the original basis curves. I have, however, assigned
"v" to the second value in patchCoords, so that if we do eventually add
support for a more-useful "u" coordinate as well, the patchCoords will be
able to be supplied as (u, v, 0, 0) rather than having to be (v, u, 0, 0)
because we happened to add support for "v" first.

(Internal change: 2334742)