Remove GfMatrix4d::Transform overload that potentially loses precision
The overload of this method that takes and returns a GfVec3f can easily cause
loss of precision when the matrix in question represents a local-to-world
transformation. Before this change, client code that transforms a local GfVec3f
to a world GfVec3d had to be careful to first cast the argument to GfVec3f, in
order to select the overload that preserves precision. This change removes the
lossy overload.

In order to support Python clients that pass GfVec3f arguments (which don't
implicitly convert to GfVec3d in Python, as they do in C++), this change adds
wrapping that acccepts GfVec3f and returns GfVec3d.

The bulk of this change is updates to call sites where the result of a call to
Transform now needs to be cast to GfVec3f. The advantage of requiring these
casts is that the loss of precision is explicit.

(Internal change: 2334853)