Fixing a bug with Pcp dependencies and change processing where some prim indexes that are the result of multiple nested relocates would not be invalidated when one of the ancestral relocates is updated.
The issue can occur when the prim index graph results in a chain of
consecutive nodes that are all relocates nodes. When dependency tracking
maps node spec paths to prim index paths, it makes a special case for
for relocates nodes where the path has to be manually mapped to the
parent before mapping to the root because relocates nodes themselves do
not contain the relocation in the map function (the function is always
identity). However, this special casing assumes that the parent node's
map function can correctly map the adjusted path. But this is not the
case if the parent itself is also a relocates node as it will not
account for its own mapping. Thus, we end up with missing/incorrect
dependencies on ancestral relocates sites.

The fix is to simply extend this relocates special casing to manually
map node paths up to the first non-relocates ancestor before applying
the map function.

(Internal change: 2335039)