convert exposure scale to logarithmic exposure in adapter.
the intent of hijacking the current GetExposure() on the HdCamera was to automatically and transparently upgrade existing delegates to the extended behaviour provided by the new exposure controls. However, there was a bug in the implementation: existing delegates will be expecting the result of GetExposure() to be logarithmic, i.e. they use the result by raising 2^exposure, whereas the result of UsdGeomCamera::GetExposureScale() is a simple scalar multiplier. So, to keep backwards compatibility we need to take the log2() of the exposure scale before handing it off to Hydra