Update hd with new MaterialOverrideSchema and add docs to relevant schemas.
The schema is designed to live on a 'material' prim type as a data source of the following structure:
materialOverride.interfaceValues.<publicUIName>.value -> value

Note that 'materialOverride' is a container data source with one member 'interfaceValues'.
Each interfaceValue specifies the overriding data source, which follows the MaterialNodeParameter schema.

(Internal change: 2336028)