[hgiVulkan] One piece of Autodesk pull request #3170, a change meant to fix failing HdSt tests when using Lavapipe. However, this change fixes a hang in many situations when using HgiVulkan, not just for Lavapipe.
From PR description:

testHdStDrawItemsCache
- Problem: is consuming command buffers faster than Lavapipe can execute them,
so it reaches the maximum allowed inflight of 64. Normally buffers would be
reset by _EndFrameSync after submission (or by EndFrame if it gets called), but
since this sync is non-blocking, and Lavapipe is slow, it's called too early
and does nothing (because all the buffers are still executing). And since all
these calls must be on the main thread, it deadlocks: main thread is waiting
for more command buffers, but those require calling _EndFrameSync to be made
available, which it can't do because it's blocked...
- Fix: Be smarter about allocating in-flight bits: search for the next
available bit. If no in-flight bits are available, then update the existing
command buffers in-flight status to release ones no longer in use, instead of
just waiting (which is a deadlock).

See #3170.

(Internal change: 2336072)