python: Initial import of boost::python from boost 1.85.0
This change just brings boost::python into the tree as a starting
point for the pxr_boost::python implementation. The src/ and test/
directories are placed next to the public headers since we don't
have separate public/private directory hierarchies like boost,
but the code itself is unchanged and not included in the build.

We use boost 1.85.0 as the starting point even though we're
currently on VFX Reference Platform 2022, which specifies
boost 1.76.0. This is because later versions of boost are
needed for other platforms and versions of Python that
OpenUSD supports. For example Python 3.11 requires boost
1.82.0+.

(Internal change: 2334040)
(Internal change: 2336880)