[hdGp] Don't cook procedurals during scene queries.
Fixes a subtle bug when GetChildPrimPaths(procPrimPath) is called before the added notices for the procedural are sent/processed. In this scenario, the added notices for generated prims weren't being sent at all (even though the procedural was being cooked again in response to the added notice) causing the generated prims to not show up.

It also makes the behavior in GetPrim and GetChildPrimPaths consistent in that we don't cook the procedural in either of these scene query methods. When queried with an uncooked procedural prim path or a not-yet-generated child prim path (e.g. notice-less traversal), we simply return the input scene's result.

(Internal change: 2337499)