python: Workaround distcc build issues with boost
distcc's pump mode fails for headers that use file iteration via
BOOST_PP_ITERATE when the iterated file is different than the
iterating file where BOOST_PP_ITERATE is invoked. In this case,
distcc fails to send the iterated file to the build machine,
leading to "File not found" errors during compilation. Per
https://github.com/distcc/distcc/issues/16 this appears to be a
long-standing known issue.

boost::python has two headers that run into this issue. As a
workaround, we add a direct include of the iterated file
surrounded by a (helpfully pre-existing) define that elides
the contents of that file. This gets distcc to recognize that
it needs to send the iterated file to the build machine without
having any other effect.

This is a temporary workaround; future changes will remove the
use of BOOST_PP_ITERATE altogether.

(Internal change: 2337543)