python: Accommodate BOOST_PYTHON_NO_PY_SIGNATURES in tests
Update tests that compare the output of help(...) to explicitly
enable Python docstrings, so they don't fail when built with
BOOST_PYTHON_NO_PY_SIGNATURES defined.

This is primarily needed because the cmake build globally
defines this symbol in cmake/defaults/CXXDefaults.cmake and
there isn't a clean way to undefine it for certain libraries.
We should be able to remove that global define when we start
using our own Python binding library throughout OpenUSD.
Because of that, it seemed better to make these tests resilient
to this symbol being defined instead of trying to untangle
the build.

(Internal change: 2337544)