Fixing a bug where ancestral relocates that should be elided were not being elided which can result in tripped verifies, crashes, and infinite loops when an ancestral variant selection is involved.
The issue was partially caused by the fact that we could still process
ancestral relocates arcs on nodes that have been elided because
of another relocates arc. This was because originally ancestral node
subtrees without specs would be marked for culling immediately but we'd
still have to check them for relocates that would "uncull" them since
relocates don't appear as specs on the site itself. So we'd always
process relocates on any ancestral node regardless of whether it was
culled or not resulting in sometimes processing relocates on nodes that
have been elided.

But changes were made to prim index node culling in 24.05 such that the
culling of nodes for lack of specs was changed to happen at the end of
the prim index building instead of during, making the additional
allowance for processing relocates on "marked as culled" ancestral nodes
unnecessary.

Additionally in 24.05, we changed how we process ancestral variant
selections for computed subtree arcs (of which relocates arcs are always
computed subtree arcs) such that the variants are delayed to be
processed after the subtree is added. This combined with the fact that
the culling behavior no longer goes back to uncull ancestor nodes when a
new arc is added and that we could still process culled relocates arcs,
led to cases where an unculled ancestral variant arc could be added
under a relocates under a culled node. And this ends up tripping a
TF_VERIFY in prim index finalization (and later a crash or hang if the a
child of that prim index is computed) where we expect to never find an
unculled node under a culled node.

With all this said the fix is to just never evaluate relocates on culled
nodes since we now know they will always be truly culled (and not just
ancestral nodes with no specs).

I've added a museum test case that demonstrates the correct behavior
that would fail due to the TF_VERIFY without the change to primIndex.

(Internal change: 2337691)