Fix divide by zero error in MaterialX Shader
MaterialX lights have a different struct and only define lights in terms
of color and intensity instead of having ambient, diffuse and specular
colors. These colors that we get in the shader essentially are pre-multiplied
by the intensity color so we approximate the intensity with the max value
of the diffuse color channels, and adjust the light color by dividing the
diffuse color by that approximated intensity value. This breaks down when
there is one light that only
has ambient light meaning we calculate the intensity to be 0 and then the
light as being 0/0.. As a result the entire object with this material is black.

This change guards the calculation of the adjusted light color to make sure
we do not divide by zero.

(Internal change: 2337962)