Attempt to fix an issue with the MergingSceneIndex.
When triggering a "whole scene" removal, this scene index can cause the
whole scene to be "added", and nothing to be "removed", which is a
problem. This can be triggered by the removeal of a sublayer from the
stage. This way of handling the change causes accesses to expired
UsdPrims, which throw exceptions.

This commit attempts to be more precise about the add and remove changes
sent in this situation by "adding" prims that are still in any merged scene
index, and "removing" any prims that exist only in the scene index that
triggered the change. This can result in a _lot_ of small changes instead
of one big change, so there may be a need to optimize this, but I think the
principle is correct (and the commit does prevent the invlid UsdPrim access
exceptions).