GeomUtilCapsuleMeshGenerator is incorrect when radii differ
This change adjusts the center and radius of the sphere caps for the capsule
mesh generation so that the height, bottom radius and top radius of the
cylindrical portion of the capsule is retained and the spheres are adjusted
to maintain continuity with the cylinder sides.

This change also alters how the numCapAxial parameter works when the radii
are different. When they radii differ, the numCapAxial parameter is doubled
and the number of cross-sections are divided between the top and bottom
hemispheres relative to the angle that each portion uses. This makes the
density of the mesh more even between the bottom and top when the radii
difference is large.

(Internal change: 2339669)