[hd,hdx,hdSt] Removed GS selection processing
The color of selected wireframe mesh edges is now resolved
independent of the rasterization order of adjacent mesh faces
so it is no longer necessary to offset selected mesh edges
in the geometry shader.

This is achieved by extending the selection buffer data to
include an encoding of the incident edges of all elements (faces)
adjacent to a selected face such that these edges can be drawn
using the correct selection color.

This needs to be distinct from other selected edges since these
edges are treated as selected elements when applying selection
colors.

The indices of these additional incident edges are obtained using
HdMeshEdgeIndexTable.

(Internal change: 2340452)