[hdSt] Fixed PTVS vertex and varying interp for triangulated quads
The PTVS shaders process a triangulated quad as a six vertex patch
representing the vertices of the original quad split into two triangles
with a vertex sequence of v0,v1,v2, v2,v3,v0. To correctly interpolate
vertex and varying primvar we need to sample primavar values at the
corners of the quad which correspond to the triangulated quad vertices
with indices: 0,1,2,4.

Also, fixed the order in which vertices are passed to the PTVS
quad and tri-quad primvar interpolation methods.

Fixes #3248

(Internal change: 2340457)