[usdImagingGL] Updated tests for vertex and varying displacement
Added tests for vertex and varying texture coordinates being used
for displacement along with updated baseline images for faceVarying.

The handling of primvar data used to access displacement can be
different than the handling of primvar data used to access other
material parameters because displacement is executed upstream of
rasterization, i.e. in the GS or PTVS shader stages while other
material parameters are typically accessed in the FS.

Also, adjusted the lighting used in the tests so that the baseline
images are not as dark as they have been previously.

Fixes #3248

(Internal change: 2340461)