python: Fix crash with overloads with keyword args and default values
Certain combinations of the above would cause segfaults when one
one of the overloads was invoked with a keyword arg. One such example
has been added to the "keywords" test.

This is a bug in upstream boost::python. Pixar has been using this
patch to fix this in internal builds since 2019, if not earlier.
This bug and fix was reported to the boost::python maintainers
but was never merged.

This patch was originally written by Alex Mohr (@gitamohr).

(Internal change: 2340606)