Extend GeomUtils to Generate Normals and Prevent Shrinkage
This change adds normals generation to the varous mesh generation utility
classes.  Similar to the point generation, the classes provide a method that
return the number of normals and a method for generating the normals into
a provided container iterator.  In addition, the interpolation mode of the
normals can also be requested.

(Internal change: 2341219)
(Internal change: 2341220)
(Internal change: 2341240)