python: Remove boost::preprocessor use for type_list
As the name implies, type_list is just a simple container of types
for metaprogramming. boost::preprocessor was used to set up the
type_list struct as well as for clients that used type_lists
internally. This has all been replaced with variadic templates.

type_list previously derived from mpl::vectorN, which is also
a simple list of up to N types, where N would be computed based on
a preprocessor define. This allows type_list to be used with the
whole suite of boost::mpl utilities. type_list now derives from
mpl::vector instead of vectorN, so it can still be used with the
boost::mpl utilities, but doesn't require extra code to figure
out what vectorN to use. This simplifies future changes.

vectorN requires less work to instantiate than vector, which is
more general. This ultimately doesn't really matter to us, since
we're going to stop using mpl::vector in a later change anyway.

(Internal change: 2341414)