Fix leaking Vulkan buffer from HdSt_PipelineDrawBatch
This went undetected by the validation layer because the
tests were also leaking the Vulkan instance and device,
so they were never destroyed, and the validation check never
kicked in. So fix the leaking tests too.

This also fixes an occasional crash from a driver trying
to use a resource after the tests had exited main(),
because the HgiVulkan leak also meant no synchronization.