Fix issue with Storm and rendering custom nodes that use texture coordinates
This issue is specific to custom nodes that use the <texcoord> node in their
defining nodegraph and either re-use that nodes output or are a multioutput
node.

This change removes the addition of the 'index' input to the network
interface, and in Storm sets the texture coordinate primvar type as
Vector2 regardless of the node's output type.

Previously, we were adding the 'index' input to the network interface which
means that the second time we come across this node we have an input
that is not a part of the nodedef. This means that when we pass it to
MaterialX for shadergen it is not able to correctly identify the node. Also,
in Storm when we gather the primvar information for the glslfx header,
we use the node's output type for the texture coordinates, which is not
likely to be a vector2 type, and for multioutput nodes specifically it
causes shadergen to completely fail.

(Internal change: 2341878)