This change contains modifications to how PCP generates changelists for layer muting / unmuting and adding / removing sublayers.
The new behavior results in PCP generating a minimal set of changes for layer operations which results in a performance increase when working with large stages.  Building upon work done in change 2336797, the SdfLayer::CreateDiff method is now used in conjuction with with PcpChanges::DidChange to compute a more percise set of change entries compared to the previous 'big blast' invalidation of '/' which resulted in a recomputation of the entire stage.

Additional notes relevant to this change:
- Change notifications to external clients remain unaffected by simulating the old behavior.
- When processing layers that may contain reloacates, the legacy behavior is used.
- This new functionality can be controled for all cases using the PCP_ENABLE_MINIMAL_CHANGES_FOR_LAYER_OPERATIONS environment setting.

(Internal change: 2342828)