build_usd.py: Remove boost dependency when building Python modules
OpenUSD now uses its internal pxr_boost::python library for building
Python modules, so boost is no longer needed. This was the last use
of boost, which means build_usd.py no longer needs to download and
build boost at all when building the OpenUSD core.

As a temporary backwards compatibility and testing mechanism, clients
can use the "--boost-python" option to revert to the previous behavior
of using boost::python for Python modules. This option is deprecated
and will be removed in a future release.

(Internal change: 2343201)