Added dependentNamespaceEditUtils to Pcp to provide a functionality for getting a description of all edits that need to be performed to fix up composition dependencies in Pcp caches when a namespace edit is performed using UsdNamespaceEditor.
The interface to this functionality is through the struct
PcpDependentNamespaceEdits which holds the descriptions of edits to perform
and the function PcpGatherDependentNamespaceEdits which computes and returns
this struct. The input to this function is the old prim spec path, the new
prim spec path, the layers in which we've already determined that the original
edits from old path to new path will take place (usually all in the same
primary edited layer stack), and a list of PcpCaches we want to find prim
index dependencies and dependent edits for.

Tests for this new functionality are not included in this change and will not
be part of Pcp. Instead a thorough suite of tests will be provided for the
UsdNamespaceEditor when it is updated to use this functionality.

This change also contains some API additions in PcpNode, PcpPrimIndex, and the
PcpLayerRelocatesEditBuilder that are needed to facilitate the dependent edit
gathering as well as a small fix to PcpNodeIntroducesDependency with respect
to inert pre-propagation specializes nodes.

(Internal change: 2343428)