This change provides additional performance improvements to the fined grained PCP changelists that were added in 2342828.
When processing changes for large stages, the PcpChanges::_didChangeSpecsInternal set could grow very large.  A performance gain can be realized by performing a more restrive dependency search and storing only topmost specs that changed in a new set.  When applying changes to the cache, the prim index is used to determine the full set of specs to update.

Additionally when muting and unmuting layers, we also perform a more restrive dependency search by only recusing on site paths.

(Internal change: 2343722)