Fix ComputeExtent for UsdGeomPointInstancer with prototypes under an over
Fixes GH-3395. As described at
https://openusd.org/docs/api/class_usd_geom_point_instancer.html#UsdGeomPointInstancer_protoProcessing
PointInstancers can define prototypes under a parent prim specified as an over.
The ComputeExtent plugin registered for PointInstancer was not correctly computing
prototype bboxes in this case. This extends UsdGeomBBoxCache to allow for a custom
prim predicate to be passed, which the updated ComputeExtent plugin takes advantage of
to traverse prototype prims in this case.