build_usd.py: add required libraries in dependency order
When libraries are added to requiredDependencies, they will be built
in the order in which they are appended to the list. This means that
any intermediary dependencies should be appended before the "primary"
library. For example, if Ptex is enabled, zlib must be built first, so
ZLIB must be appended to requiredDependencies before PTEX.