Add support for man pages to CLI tools
This PR adds support for adding man pages to the command line utilities.

As a sample, I've translated just usdcat over from https://openusd.org/release/toolset.html  to a manpage.

It is possible to use pandoc to generate both the manpage and the sphinx doc, but I wanted to keep this PR small.

While Windows does not technically use manpages, there are various systems on windows that allow accessing them, so I haven't gated these behind a platform check.