APPLE: Remove multiple instances of ZLIB
Due to a prior change in OpenUSD, ZLIB was moved to a conditional dependency, and then removed from the list if run on Linux or macOS

Unfortunately, the removal only removes the first found instance of ZLib, so doesn't remove it if multiple dependencies ask for it.

A future clang version also has issues with the ZLib version used here so this would be needed to pass builds soon. However, since clang is used primarily on Linux and Mac, it is preferred to just use the OS provided version of Zlib.