Add Hgi functions to get and set default backend
This may be used by the host application to set the preferred backend, and
then plugins can follow it.

It also works around an issue where Storm calls Hgi::IsSupported without
specifying a backend, which may fail when the host application uses Vulkan
and the default OpenGL backend does not work.