Move MATERIALX_GENERATE_INDEX macro
Moving MATERIALX_GENERATE_INDEX and MATERIALX_VERSION_INDEX above the MaterialX ShaderGenerator includes as there will likely be another include conditionally added for WGSL based on the MaterialX version.  Putting the code in the same place so it does not have to be moved later and create merging confusion.

Tested with MaterialX 1.39.3 and 1.38.10 versions on MacOS.