Fix 3666: Crash when using Vulkan + AMD
AMD driver (both Mesa + Pro) have limits that cannot fit inside an int.
This resulted in sending incorrect vulkan commands as the number of work
send to the GPU would be set to the max limit in stead of the max amount
of work that needed to be done.

A better solution would be to introduce `GfVec3u`.

This has been validated to work on Windows/Linux with several GPU
generations (RDNA3.5, RDNA4, RDNA3).