[hgi] On shader compiling errors, output ALL source with line numbers.
The normal hgi behaviour is to log just the reported errors.
To make these errors faster to deal with, the shader compiling logs the full source, annotated with the same line numbers as the driver error messages.

This output will only be seen if there is a shader compiling error.

hgi GL, Vk, and Metal have been patched to do the same thing.

Fixes #4048