APPLE: Make Accessibility Priority variable
We have been showing the USD Accessibility API to more members of the accessibility community and a request was made to make priority not be time varying. Originally, we made this uniform because we were thinking of it as a way to help prioritize accessibility purposes globally, but have realized that this would also benefit from being varying.

For example, if there was an accessibility purpose that was only high priority for a given part of an animation, like flashing red lights, that would be useful for a system to know when it changes.

Looking at the USD code and spec, this should be a change that doesn't necessitate versioning the schema. Coupled with the low likelihood of anything having adopted the schema yet, and the even lower likelihood of people using the priority attribute yet, I think this is a safe assumption. I've done a few tests to make sure that nothing in USD seems to get mad when the uniform token is taken away or added as well, and all seems fine.