Add units selection for UsdGeomCamera
The 'tenths of a scene unit' behavior for camera geometry causes no end of confusion and frustration in mixed-unit pipelines, as well as being a source of bugs in implementations.

This PR adds a  attribute to UsdGeomCamera to allow selecting between 'sceneTenths' (the default) and 'millimeters'. The latter specifies that all camera geometry (i.e. aperture, focal length) is specified in millimeters, and that renderers must handle conversion from millimeters to stage units themselves.